requirements.txt
flask
mysql-connector-python
requests
python-dotenv